[IMP] im_livechat: introduce several PublicLivechatGlobal constants
Task-2899918
